+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.1.0
+
+- Big refactor of the codebase.
+- Can generate models from json file.
+- Can generate controllers from json file.
+- Can generate services from json file.
+
+## 1.0.4
+
+- Big refactor of the codebase.
+
 ## 1.0.2
 
 - Fixed issue with pubspec.yaml file.